+++ routes/quest-creation.js
@@ -1,6 +1,6 @@
 var express = require('express');
 var router = express.Router();
-var uploadPhoto = require('../controllers/upload-photo.js');
+var uploadPhoto = require('../controllers/quest-creation.js');
 var multer = require('multer');
 var upload = multer({ dest: './uploads/'});
 var loginRequired = require('../middlewares/auth.js').loginRequired;
